Add prefix withXxxYyyy to mixins, keep file name as XxxYyy Rationale: the module path/folder makes it clear that it refers to a mixin (and were kept that way), however when one is composing or consulting the subclass, it is useful to promptly know which one is a mixin, because it helps to dismiss worries about "strong"* multiple inheritance, which is not the case with mixins. "strong"* -> inheritance involving real/complete entities that can live by themselves.
ps. I opted by lowercase at the start, because 'with' is not part of the effective name of the mixin.
Feel free to change it if it clashes with other principles. [davi]

** -> I consider this case an exception not thought when the linter rules were created.
E.g. 'list', 'map' etc. are also classes with lower case and are so for mere typing convenience AFAIK (a case even weaker than semantics).